Move notebook debugger into notebook folder
