+++ src/platform/debugger/jupyter/helper.node.ts
@@ -3,7 +3,7 @@
 
 import { DebugProtocol } from 'vscode-debugprotocol';
 import { IKernel } from '../../../kernels/types';
-import { IKernelDebugAdapterConfig, KernelDebugMode } from '../types';
+import { IKernelDebugAdapter, IKernelDebugAdapterConfig, KernelDebugMode } from '../types';
 
 export enum IpykernelCheckResult {
     Unknown,
@@ -121,4 +121,14 @@ export function isShortNamePath(path: string): boolean {
 export function shortNameMatchesLongName(shortNamePath: string, longNamePath: string): boolean {
     const r = new RegExp(shortNamePath.replace(/\\/g, '\\\\').replace(/~\d+\\\\/g, '[^\\\\]+\\\\'), 'i');
     return r.test(longNamePath);
+}
+
+
+export async function cellDebugSetup(kernel: IKernel, debugAdapter: IKernelDebugAdapter): Promise<void> {
+    // remove this if when https://github.com/microsoft/debugpy/issues/706 is fixed and ipykernel ships it
+    // executing this code restarts debugpy and fixes https://github.com/microsoft/vscode-jupyter/issues/7251
+    const code = 'import debugpy\ndebugpy.debug_this_thread()';
+    await kernel.executeHidden(code);
+
+    await debugAdapter.dumpAllCells();
 }
