+++ src/platform/debugger/jupyter/notebook/debuggingManager.node.ts
@@ -17,25 +17,26 @@ import {
     EventEmitter,
     NotebookEditor
 } from 'vscode';
-import * as path from '../../../platform/vscode-path/path';
-import { IKernel, IKernelProvider } from '../../../kernels/types';
-import { IConfigurationService, IDisposable } from '../../common/types';
-import { KernelDebugAdapter } from './kernelDebugAdapter.node';
-import { IExtensionSingleActivationService } from '../../activation/types';
-import { ContextKey } from '../../common/contextKey';
-import { IApplicationShell, ICommandManager, IVSCodeNotebook } from '../../common/application/types';
-import { traceError, traceInfo, traceInfoIfCI } from '../../logging';
-import { DataScience } from '../../common/utils/localize';
-import { Commands as DSCommands, EditorContexts } from '../../../webviews/webview-side/common/constants';
-import { IPlatformService } from '../../common/platform/types';
-import { IDebuggingManager, IKernelDebugAdapterConfig, KernelDebugMode } from '../types';
-import { DebuggingTelemetry, pythonKernelDebugAdapter } from '../constants';
-import { sendTelemetryEvent } from '../../../telemetry';
-import { DebugCellController, RunByLineController } from './debugControllers.node';
-import { assertIsDebugConfig, IpykernelCheckResult, isUsingIpykernel6OrLater } from './helper.node';
-import { Debugger } from './debugger.node';
-import { INotebookControllerManager } from '../../../notebooks/types';
-import { IFileSystem } from '../../common/platform/types.node';
+import * as path from '../../../vscode-path/path';
+import { IKernel, IKernelProvider } from '../../../../kernels/types';
+import { IConfigurationService, IDisposable } from '../../../common/types';
+import { KernelDebugAdapter } from '../kernelDebugAdapter.node';
+import { IExtensionSingleActivationService } from '../../../activation/types';
+import { ContextKey } from '../../../common/contextKey';
+import { IApplicationShell, ICommandManager, IVSCodeNotebook } from '../../../common/application/types';
+import { traceError, traceInfo, traceInfoIfCI } from '../../../logging';
+import { DataScience } from '../../../common/utils/localize';
+import { Commands as DSCommands, EditorContexts } from '../../../../webviews/webview-side/common/constants';
+import { IPlatformService } from '../../../common/platform/types';
+import { IDebuggingManager, IKernelDebugAdapterConfig, KernelDebugMode } from '../../types';
+import { DebuggingTelemetry, pythonKernelDebugAdapter } from '../../constants';
+import { sendTelemetryEvent } from '../../../../telemetry';
+import { DebugCellController } from './debugCellControllers.node';
+import { assertIsDebugConfig, IpykernelCheckResult, isUsingIpykernel6OrLater } from '../helper.node';
+import { Debugger } from '../debugger.node';
+import { INotebookControllerManager } from '../../../../notebooks/types';
+import { IFileSystem } from '../../../common/platform/types.node';
+import { RunByLineController } from './runByLineController.node';
 
 /**
  * The DebuggingManager maintains the mapping between notebook documents and debug sessions.
